認可機能実装（サイドバー）
