--- app/Http/Controllers/Api/AuthController.php
@@ -1,91 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\Api;
-
-use App\Models\User;
-use Illuminate\Http\Request;
-use App\Http\Controllers\Controller;
-use Illuminate\Support\Facades\Hash;
-use Illuminate\Validation\Rules;
-use Illuminate\Support\Facades\Validator;
-
-class AuthController extends Controller
-{
-    public function register(Request $request)
-    {
-        $validator = Validator::make($request->all(), [
-            'name' => 'required|string|max:255',
-            'email' => 'required|string|lowercase|email|max:255|unique:'.User::class,
-            'password' => ['required', 'confirmed', Rules\Password::defaults()],
-        ]);
-
-        if ($validator->fails()) {
-            return response()->json([
-                'validation_errors' => $validator->messages(),
-            ]);
-        } else {
-            $user = User::create([
-                'name' => $request->name,
-                'email' => $request->email,
-                'password' => Hash::make($request->password),
-            ]);
-
-            $token = $user->createToken($user->email . '_Token')->plainTextToken;
-
-            return response()->json([
-                'status' => 200,
-                'username' => $user->name,
-                'token' => $token,
-                'message' => 'ユーザー登録が成功しました',
-            ]);
-        }
-    }
-    
-    public function login(Request $request)
-    {
-        $validator = Validator::make($request->all(), [
-            'email' => 'required|string|max:255',
-            'password' => 'required']);
-
-        if ($validator->fails()) {
-            return response()->json([
-                'validation_errors' => $validator->messages(),
-            ]);
-        } else {
-            $user = User::where('email', $request->email)->first();
-
-            if (! $user || ! Hash::check($request->password, $user->password)) {
-                return response()->json([
-                    'status' => 401,
-                    'message' => '認証に失敗しました',
-                ]);
-            } else {
-                $token = $user->createToken($user->email . '_Token')->plainTextToken;
-
-                return response()->json([
-                    'status' => 200,
-                    'username' => $user->name,
-                    'token' => $token,
-                    'message' => 'ログインに成功しました',
-                ]);
-            }
-        }
-    }
-
-    public function logout(Request $request)
-    {
-        if ($user = auth()->user()) {
-            $user->tokens()->delete();
-            return response()->json([
-                'status' => 200,
-                'message' => 'ログアウトしました'
-            ]);
-        } else {
-            return response()->json([
-                'status' => 401,
-                'message' => '認証情報がありません'
-            ]);
-        }
-        
-    }
-}
