+++ resources/js/types/index.d.ts
@@ -1,12 +1,15 @@
 export interface User {
-    id: number;
-    name: string;
-    email: string;
-    email_verified_at: string;
+    id: number
+    name: string
+    email: string
+    is_admin: number
+    email_verified_at: string
 }
 
-export type PageProps<T extends Record<string, unknown> = Record<string, unknown>> = T & {
+export type PageProps<
+    T extends Record<string, unknown> = Record<string, unknown>
+> = T & {
     auth: {
-        user: User;
-    };
-};
+        user: User
+    }
+}
